Include color mode (xy or colortemp) for hue rgbw bulbs
The color mode can be used by home automation to determine the correct way to (re)calculate the set lighting. Zigbee2mqtt plugin for Domoticz will use the acknowledged info to correctly set the sliders (rgb/colortemp) in the UI and database. (my upcoming addition to the plugin to support HUE white and ambiance bulbs will use it)
The hue bulb will only update this value after a color mode change, but it will be cached by MQTT.