Added light blinking (Hue style 'alert') feature
The feature used is not specific to Philips Hue, but your milage may vary. Tested on Philips Hue White and color ambiance and on Ikea Tradfri White.

Publish the following commands to your lightbulb thru mqtt:
{"alert":"select"} for a single blink
{"alert":"lselect"} for a longer blink time (I get 15 blinks in 15 seconds on Philips Hue)
{"alert":"none"} stop blinking 

The bulb will flash in the color you have set.
If the bulb is off you should first turn it on and set the desired color.