Add combined toZigbee color and colortemp converter. (#260)
* Add support for AduroSmart ERIA White and Color bulbs.

* Update devices.js

* Add combined toZigbee color and colortemp convertor.

This convertor is a combination of light_color and light_colortemp and
can be used instead of the two individual convertors. When used to set,
it actually calls out to light_color or light_colortemp to get the
return value. When used to get, it gets both color and colorTemp in
one call.
The reason for the existence of this somewhat peculiar converter is
that some lights don't report their state when changed. To fix this,
we query the state after we set it. We want to query color and colorTemp
both when setting either, because both change when setting one. This
converter is used to do just that.

We're using this convertor by default now in
generic.light_onoff_brightness_colortemp_colorxy because it can't hurt.

* Typos.