Add device to support Trust Remote control ZYCT-202 (#300)
* Add device to support Trust Remote control ZYCT-202

https://github.com/Koenkk/zigbee2mqtt/issues/635

* Add the concerters for the Trust Remote control ZYCT-202

The groupid has been processed in the button press. So you have the action and the selector mode in 1 state value and can easily be used in a Home Assistant Automation state rule.

* Add the concerters for the Trust Remote control ZYCT-202

The groupid has been processed in the button press. So you have the action and the selector mode in 1 state value and can easily be used in a Home Assistant Automation state rule.

(https://github.com/Koenkk/zigbee2mqtt/issues/635)

* Update devices.js

Fixed spaces line 1941

* Update devices.js

Fixed the maximum line length of 120  max-len

* Separated the action and groupid in the return rules

Have separated the action and groupid in the return rules...

* Update fromZigbee.js