Add support for Heiman HS3MS, HS3WL, HS3DS
Hi

I'm new to this and spent a lot of time adding these devices.  All devices are under //HEIMAN section. They all work with the alterations to this file and the converter(s).  The only issue I had was the HS3WL, HS3DS coming up as ModelID 'undefined' when adding?  HS3WL, HS3DS use the same converters as the HS1WL and HS1DS models.

Regards
Quinn