fromZigbee should handle currentHue in addition to currentSaturation (#773)
After this change we properly handle bulbs that do standard currentHue + currentSaturation,
according to the spec both should always be published.

```
5.2.2.2.1.1 CurrentHue Attribute

The CurrentHue attribute contains the current hue value of the light. It is updated
as fast as practical during commands that change the hue.
The hue in degrees shall be related to the CurrentHue attribute by the relationship
Hue = CurrentHue x 360 / 254 (CurrentHue in the range 0 - 254 inclusive)
If this attribute is implemented then the CurrentSaturation and ColorMode
attributes shall also be implemented.

5.2.2.2.1.2 CurrentSaturation Attribute

The CurrentSaturation attribute holds the current saturation value of the light. It is
updated as fast as practical during commands that change the saturation.
The saturation shall be related to the CurrentSaturation attribute by the
relationship
Saturation = CurrentSaturation/254 (CurrentSaturation in the range 0 - 254 inclusive)
If this attribute is implemented then the CurrentHue and ColorMode attributes
shall also be implemented.
```

In addition to handle currentHue, currentSaturation got update to round to the nearest whole number.

Using https://www.rapidtables.com/convert/color/hsv-to-rgb.html we can now convert the resulting values back, having V=100 set we get the color we configured.

This was tested with an Innr RB 250 C.