fix: SP600 spamming (#1157)
Modifies the `instantaneousDemand` and `currentSummDelivered` reporting
configuration to:

* Minimum interval of 5 seconds
* Reportable change of 1 W (given the current divisor of `10,000`)

`currentSummDelivered` is a `uint48` value and as such, when reporting
is being configured, the `reportableChange` needs to be of the same
type. `BuffaloZcl.writeUInt48` expects to be passed a 2 element array of
numbers, the first representing the 2 most significant bytes and the
second the 4 least significant bytes. This means the default payload had
an incorrect value of `1`, which should instead be `[0, 1]`.

Fixes #903