Update Zen thermostat
This change enables reporting of extra parameters for the Zen thermostat.

I'm also changing the "operation" parameter on some other thermostats to
align with running state (as per the thermostatHeatCool templates that are
applied to a few other thermostats in the home assistant auto config logic)

This change will be matched by a commit to the HA autoconfig code such that
the thermostats which used the "operation" field should effectively see a
no-op.