Use float types for tuya temperature readings (#1805)
As I was visualizing some of the logged data, I found that the temperature readings are being sent out as strings instead of floats. I guess this was done because of the toFixed that truncates the number of decimals (and then turns it into a string). However most of the tooling for visualization expects this data to be numbers and not strings.