scene_add should handle color_temp for ZCL version <= 6
ColorTemperatureMireds was added in ZCLv7, not many (any) currently support his.

```Since there is a direct relation between ColorTemperatureMireds and XY, color temperature, if supported, is stored as XY in the scenes table.```

Seems to indicate it is OK to store it as XY instead, this change introduces a new optional property `ZCL` that specifies the ZCL version.
We default it to v6 for most compatibility, but when set to >= 7 we will use the new attribute. This will also allow us to do other things in the future
if the ZCL gets updates again.

In practice the default of ZCLv6 should be fine for most people, bulb seem to properly report colorTemperature after the change.
(Tested on Trådfri WS)