Ubisys S1/S2 can support action
When the control endpoints of a S1 (only 1 endpoint) or an S2 (2 endpoints) are bound to a group they emit commands.
Before they were only captured for Ubisys C4, we now capture them for S1 and S2 too.

I don't have access to other devices to add them. I've decided to go with the standard way of implementing this instead of extending ubisys_c4_* fromZigbee converters.
As the default way of handling this works nicely.