scene_add should also accept "Invalid Field" for the remove call
During testing I would occasionally get "Invalid Field" back as a respons for the remove command.

We'll just add the new scene like we do for "Not Found". Not 100% sure why we're getting that reply.
But once we do, the scene seems to be absent from the bulb and can't be recalled anymore.

It seems save to execute the add in that case.