Moes thermostat negative local temp
Local temperature comes as a 16bit signed integer with the last digit as a fraction. When chunks are saved into JS's integer sign bit is lost. As a result, for example, at -2.9 local temperature zigbee converter outputs 6550.6 .
This change fixes this problem. I am not sure what is the logic behind  "case tuya.dataPoints.moesTempCalibration:" is, it seems to be doing something close but then it has to be -4095 and be compared to 2048.