devices: SNZB-02 and SNZB-04 should use batteryPercentageRemaining (#2347)
* devices: SNZB-02 and SNZB-04 should use batteryPercentageRemaining

Both SNZB-02 and SNZB-4 have batteryPrecentageRemaining in the database.db and seem to report it daily.
This would explain the battery % variang a bit as mentioned in koenkk/zigbee2mqtt#6657

The SNZB-02 was easy to configure most will do it on the first new message or when the button is hit.
The SNZB-04 was annoying, you need to either remove + join or remove the battery for it to  take the new reporting values. They also seem to apply when either batteryVoltage or batteryPercentageRemaining come in via attReport. But they don't apply when the ias_contact_alarm is triggered.

The configureKey for SNZB-04 was not bumpbed to not have to many errors in the logs. But even without reconfigure batteryPercentageRemaining is reported, usually shut before batteryVoltage, so it should be fine.

* devices: SNZB-04 also expose battery_voltage