YRD226/246 pin code management (#2501)
Added pin code message support for Yale YRD226/246. This is simply enabling the support which already exists in zigbee-herdsman-converters. I have tested this on my own lock (YRD246) and all works fine. It likely works with other Yale locks too, however I cannot test this so have not enabled it.