Add support for Haozee TS0601 TRV variant
This PR adds support for the Haozee Zigbee thermostat sold on Aliex[ress. It's simply a variant of the Tuya TS0601 thermostat so it was sufficient to only add the different manufacurerName (_TZE200_a4bpgplm) to the tuya.js.
With this change the device gets detected and you can interact with it just like with any other TS0601 thermostat.