Support newer Lupus 2 (fw or board revision?)
I've just added a new Lupus 2 to my zigbee network. It showed up as unsupported, which is due to the zigbeeModel being newer than what the converter supports.

My new device has zigbeeModel PRS3CH2_00.00.05.12TC as opposed to 10TC or 11TC.

This has been tested and works as expected.