tuya_smoke: Accept commandDataReport in addition to commandDataResponse
Newer(?) version of the TS0601 (_TZE200_uebojraa) smoke detector seem to use
commandDataReport instead of/in addition to commandDataResponse payloads.

Add commandDataReport to the accepted types.  While at it, change the
converter to use the new "multi-dp for loop" approach.  (A new DP #14 shows
up when doing this, but we currently don't know what it is)

Fixes https://github.com/Koenkk/zigbee2mqtt/issues/10464