Update konke.js
It's my absolute first PR and my understanding of the process is very limited, so please accept my apologies if this was not what was expected.
Since all the constants are already called out in the header, I trimmed the code to the bare minimum, i.e. the variables for the sensor.