Minor fix to color XY calculation
The color XY uint16 representation in zigbee spec is in `s15Fixed16Number` format per ICC spec (https://www.color.org/specification/ICC1v43_2010-12.pdf). It is actually a 16-bit fixed-point number. If we don't consider negative numbers, it is exactly the integer divide by 65536 instead of 65535. Minor, but better to get it according to spec.