Philips Hue OTA
Use the meethue server to request available firmware.
If multiple firmwares are returned the oldest one is applied first.

There is firmware available for the following imagetype's that
already exists in zigbee-OTA:
273, 274, 276, 277, 279, 280 and 282.
And also available (not in zigbee-OTA):
285 and 288.

It's likely other imagetype's can be found if the format of
'DeviceTypeId' for those can be found.