Fix for commit #3839
This PR fix, what problems commit #3839 introduced.

Not all devices support acFrequencyDivisor and acFrequencyMultiplier.

With the original PR, this would introduce an UNSUPPORTED_ATTRIBUTE error on for example Heiman HS2SK, and no readings could be fetched from the device.

To access the acFrequencyDivisor and acFrequencyMultiplier, the following line should be used in the device:

await reporting.readEletricalMeasurementMultiplierDivisors(endpoint, true);

Do NOT use the following line, because it will break it again:

await reporting.acFrequency(endpoint);

Co-Authored-By: corneyl <5320363+corneyl@users.noreply.github.com>