Add on_level and execute_if_off to level_config
on_level: the brightness at which the light turns on. Differs from
current_level_startup in that the latter is specifically for when
the device is connected to power. Values: "previous", 1..254.

execute_if_off: whether setting `{brightness: VALUE, state: null}`
should have an effect even while current state is off. If so and
on_level is "previous", the next time the light is turned on it will
use the new brightness; otherwise it keeps the brightness set the last
time it was on. Values: true, false.