Do not read onLevel again if already in state
The device might have forgotten it for some reason. If we assume that
the user is not modifying it with direct attribute writes, the read
should be redundant anyway.

Also, if the device does not support OnLevel at all, publish level_config
with on_level set to 'previous' in order to never attempt to read it again.

zigbee2mqtt#13570