Do not read onLevel again if already in state
The device might have forgotten it for some reason. If we assume that
the user is not modifying it with direct attribute writes, the read
should be redundant anyway.

If the device responds with an error to a request for on_level, mark
that in globalStore and always use 'previous' from then on.

zigbee2mqtt#13570