Add emergency_heating for heating to SLR2b/c
SLR2c (which I have) supports system mode emergency_heating for the heating control as well as water. The SLR2b (which I don't have) has information about using emergency_heating for controlling heating, but not water. I assume that also supports it, but the possible value for withSystemMode has not been included by mistake. This pull request adds it to both.