MOSZB-140 some minor cleanup and fixes (#4741)
* utils: postfixWithEndpointName should handle "undefined" better

e.g. the MOSZB-140 has 35 as the default endpoint, most of the config on genBasic, genPowerCfg and ssIasZone are there.
Additionally it uses seperate endpoints for temperature and illumance (40 and 39). It also has 3 (!!) msOccupancySensing endpoints (34, 40, 41), for those we do want to provide a suffix to the property name.

For those we don't really need a suffix for the propery name, having `{default: 35, 'logic_a': 34, 'logic_b': 40, 'logic_c': 41}` mostly works but tempeature now has _undefined appended.
With this change, instead of appending undefined, we just return the original proporty.

* MOSZB-140: cleanup

- drop read of batteryVoltage, reporting configuration does this
- update reporting settings to match with the suggestions from develco
  (esp illumantion was very spanny and battery draining with the defaults)

* MOSZB-140: try/catch on read of develcoAlarmOffDelay/develcoLedControl

There was a issue were configure kept failing due to timeouts on reading these.
As show in https://github.com/Koenkk/zigbee-herdsman-converters/pull/4731 the tech manual might not be 100%
 accurate. So putting them in a try/catch should at least let the configuration continue settings up everything else.