Create akuvox.js
This Sensor is a combination of a TS0201 Tuya-Sensor with Heiman datapoints. The heiman-datapoints also has changed (battery). If it will be connected to Zigbee2MQTT, it will recognized as TS0201 temperature & humidity Sensor with display. But it has no display and zigbee2mqtt will not receive any data from that devices. 
That's the reason for me to build a new converter, hopefully, that's the right way to integrate this device into zigbee2mqtt.