Add M423-9E (#4894)
* Create akuvox.js

This Sensor is a combination of a TS0201 Tuya-Sensor with Heiman datapoints. The heiman-datapoints also has changed (battery). If it will be connected to Zigbee2MQTT, it will recognized as TS0201 temperature & humidity Sensor with display. But it has no display and zigbee2mqtt will not receive any data from that devices. 
That's the reason for me to build a new converter, hopefully, that's the right way to integrate this device into zigbee2mqtt.

* Update akuvox.js

* Update akuvox.js

* Update akuvox.js

Thanks @koenkk, your tipp is running well.

* Update akuvox.js

* Update akuvox.js

The optimized code doesn't run correctly. So rollback to the original. SORRY!!

* Update akuvox.js

* Update akuvox.js

Co-authored-by: Koen Kanters <koenkanters94@gmail.com>