Clean lookups for sinope_thermostat (#5062)
I got weird behaviors if this is not formated like this. This mainly change the lookup constants.  I think this should have been in the release ( or add the 5031pr below)

The code has been taken from https://github.com/Koenkk/zigbee-herdsman-converters/pull/5031, I forgot to add it my pr