Fix for LIDL 14149505L/14149506L color handling (#5588)
* Fix stuck color state

This fixes a bug where the color in the state wasn't updated because it
was using "h" and "s" as field names. Those fields are just copies for
Home Assistant while the actual canonical field names are "hue" and
"saturation".

* Support color sync

This allows the color sync functionality to work if it's enabled in the
device settings.