Profalux covers: switch recent ones to using cluster 0x102, add MAI-ZTP20C remote
This change references remote model MAI-ZTP20C.
More importantly, it switches the control of covers to using the dedicated ClosuresWindowCovering (0x102) cluster for the covers that support it.
Sniffing Zigbee traffic of my remotes show that the remotes use this cluster to control the covers, and not 0x8 as currently implemented.
The compliance document for the Profalux motors at https://csa-iot.org/csa_product/ac-motor-for-window-coverings/ also lists that cluster and does NOT list 0x6/0x8.

So I think the correct way to control these covers is 0x102, and certainly it's the way that seems to work with my shutters (all bought in 2023).

Older revisions of these motors, which did not expose a zigbeeModel field, did not expose 0x102 either and so these use the pre-existing fingerprint to keep using cluster 0x8.