fix: Fix invalid thermostat value handling (#5892)
* Add missing hvacThermostat attributes

* Handle invalid hvacThermostat ints

* Update invalid value handling for hvacThermostat

The most important change is NOT to return 0 in case invalid value is returned.

Second change is to detect invalid values differently: according to ZCL spec rev8, the valid range for most of the hvacThermostat attributes is 0x954d – 0x7fff. The lower bound translates to -27315 (-273.15 respectivelly). Invalid value for sint16 is 0x800. The change updates the threshold for invalid value detection from -250 to -273.15.

* Don't add invalid values to the result