Profalux covers: improve model names
The recent models we know are MOT-C1Zxxy, where "xx" is the torque (06 10 20 30Nm) and y appears to be a version (C or F seen in the wild).
Update the model number accordingly, as "NSAV061" is the name of the after sales documentation file, not the proper model number.

Keep using NSAV061_v1 for the older version, since we have no confirmation of its actual model number and it doesn't appear to even report it over Zigbee (to be confirmed).
I suspect it is MOT-C1ZxxA.