Build now works regardless of setTimeout return type
For unknown (to me) reasons, I saw builds failing because the
browser-compatible version of setTimeout/clearTimeout was being
used, instead of the Node-version, resulting in errors like:

```
src/lib/ota/common.ts:405:43 - error TS2345: Argument of type 'Timer' is not assignable to parameter of type 'string | number | Timeout'.
405                             clearInterval(timer);
```

As I'm unsure of the cause and the desired behaviour, my fix
is to support both.