Add Wenzhi WZ-M100-Z as whitelabel for TuYa ZY-M100-S_1
Add Wenzhi WZ-M100-Z as whitelabel of the TuYa ZY-M100 as this is a clone
and implement the same features.

By opening the device and looking at the PCB and the chip used, the
device use the same sensor and the same ARM controller of the TuYa ZY-M100.

Only the fingerprint reported is different.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>