fix: Set color temperature range for IKEA LED1624G9 (#6640)
This light is pure RGB, it does not contain white LEDs and it only
supports XY, not color temperature. The latter is emulated by Z2M via a
conversion to XY. The color temperature range advertised by this light is
the same one as other IKEA lights, which typically have LEDs supporting
the 2200-4000K range — but that does not make sense given that all
temperatures are simulated, and the current temperature->XY conversion
code accepts any value, including out of this range, if it is submitted
via MQTT.

Given the above, this commit changes the avertised range of this light to
2000-6500K. This has the benefit of making the whole range available in
the UI of Home Assistant (compared to only via manual service calls).