feat: VOCKQJK11LM add air_quality
Did some testing, 297 aqaraOppl value seems to be the TVOC air quality value.
I noticed this one changed from 2 -> 1 a few minutes after joining when it went from 4 to 4 leafs on the display.

I did some more testing with hairspray and was able to observe all the values except 1 so I'm fairly confident.
I mapped these to the same enum we use in other places where we have an air_quality property.