feat: Add missing clusters to Xioami VOCKQJK11LM (#6840)
* feat: add quirkAddEndpointCluster and quirkSendWhenActive modernExtend

quirkAddEndpointCluster allows to add input and output clustersto endpoints.
We can use this quirk on devices what don't have all endpoints discoverable (e.g. LUMI manufacturer devices from Aqara, Xiaomi, Maji, ...)

quirkSendWhenActive sets device.defaultSendRequestWhen to active, some devices are battery powered and sleepy but do not have the genPollCtrl cluster.
e.g. VOCKQJK11LM will timeout nearly all actions as it defaults to immediate without this quirk.

* feat: VOCKQJK11LM add inputCluster quirks

* fix: drop now undeeded endpointID override

* fix: aqaraAirQuality can briefly be 0 after startup

Imediatly after startup the airQuality attribute seems to be 0, this is displayed as `---` on the device when showing ppb value or as the complete absense of the leaves when displaying the 5 leaf mode.

* feat: want device.save() mock for tests

* tests

* fix tests

---------

Co-authored-by: koenkk <koenkanters94@gmail.com>