Partially add Schneider Electric AC Fan module support
This patch includes the fan module by Schneider electric in a partial
form. All of the practical use-cases of the fan are here and working:
- turning on/off
- setting speeds to low, medium or high

What is not included is controlling the LED, this requires a little bit
more effort and packet sniffing as the standard endpoint of the other
Wiser switches do not work for this particular model and/or firmware.