feat: add xiaomiOutageCountRestoreBindReporting extend for Aqara TVOC
This sensor loses binding and reporting information on a power cycle.

Sadly at least this device does not trigger a deviceAnnounce and will just start sending messages again :(

We can detect when this happens if the power_outage_memory increases, so when this happens we restore both bindings and reporting.