Revert referencing the attribute name
This was changed in https://github.com/Koenkk/zigbee-herdsman-converters/pull/7139
and seemed to break this integration. Unsure as to why this is not
working anymore as everything logically looks correct.

For now lets revert this change and hope this resolves the bug and gets
this part of the integration working again.