fix: Fix the broken LED indicator for Schneider Electric wiser devices (#7152)
* Fix LED indicator access on different endpoint

A previous PR broke the functionality of the LED indicator because an
assumption was made that these switches had this endpoint on the main
one as the switch. This is untrue and is located on a different
endpoint/cluster.

* Power-on behaviour not compatible with thess devices

This was added with an assumption that it would work with the device,
it's not compatible.