feat(add): TS000F based TYZGTH4CH-D1RF
This adds support for the Mumubiz TYZGTH4CH-D1RF
which is based on a Tuya ZT3L module and reports as TS000F.

The device has a connector that supports an external DHT-22
or alike one-wire temperature and humidity senor.
It might also have detection for DS18B20 sensors, but I can't verify it.