feat(add): SNZB-01-KF (#7888)
* Updated sonoff.ts to support a rare version of SNZB-01 button

Adding support to a rare version of SNZB-01, model KF01, that I own. Shares the same look as WB01 but only supports single click. Uses the "ssIasZone" endpoint to report button presses so the WB01 converter doesn't work

* Update sonoff.ts

Adding support to a rare version of the SNZB-01 button, model KF01, that I own. It shares the same look as WB01 but only supports single click. It uses the "ssIasZone" endpoint to report button presses so copying the WB01 description didn't work. Now the button works.

* Update sonoff.ts

* Update sonoff.ts

---------

Co-authored-by: Koen Kanters <koenkanters94@gmail.com>