feat(add): FanBee device
This is a custom self-built device for ventilating spaces. It contains a
valve and a speed-controlled fan, the valve is automatically opened when
the fan runs. The Zigbee standard fan cluster is unsuitable for this
type of device, so instead a level-controlled device type is used,
allowing for fine speed control over the fan.

I'll publish the design files and firmware for this eventually, but we
need one device to test the speed-controlled fan device model in Z2M so
we need it in there already.