feat: Add occupancy timeout to modernExtend.iasZoneAlarm https://github.com/Koenkk/zigbee2mqtt/issues/24621
This kind of timeout is intended for devices that periodically
send occupation notifications for as long as occupancy is detected,
while additionally sending a clearance notification when no more
motion detected. In case the clearance notification gets lost, we
can still infer that no more motion is detected if the occupancy
messages stop coming in for a while.