Fix CO2 value interpretation
The zigbee standard specifies that the values of section 4.13
Concentration Measurement are to be represented between 0 and 1 (section
4.13.2.1 Attributes). Further it is explained in section 4.13.2.1.1
MeasuredValue Attribute that "MeasuredValue represents the concentration
as a fraction of 1 (one)."

Because the current implementation multiplies the given MeasuredValue by
1000000 (million) the published value is not only wrong, but changes in
the inverse direction. The larger the actual measured value becomes (the
larger the denominator is), the smaller MeasuredValue becomes.

| Device (PPM) | MeasuredValue | z2m (PPM) |
|--------------|---------------|-----------|
| 800          | 0.00125       | 1250      |
| 1250         | 0.00080       | 800       |