Add valve flow rate to AVATTO TRV06.
This is as simple as wiring up data point 101; the Smart Life app
calls this “flow rate” (IIRC) but it seems to match well to the
concept “PI Heating Demand” that zigbee-herdsman-converters already
has for other devices.

I haven't checked if this would be somehow settable; the app shows
it as a read-only value so I've left it read-only here as well.

Tested with _TZE284_o3x45p96.