Busch-Jaeger 6735/6736/6737: Fix TABLE_FULL errors during device config
These devices sometimes fail to set up all of their bindings after pairing.
The reason for this are `TABLE_FULL` errors, which are reported by the
device. Apparently these devices only support 4 or 5 bindings, possibly
depending on the network size.

Binding the light endpoint to the coordinator can be skipped because
the BJ devices do not support reporting (they are polled instead). This
allows us to save some precious space in the binding table.

See Koenkk/zigbee-herdsman-converters#7939