Use postfixWithEndpointName in color_colortemp
Unlike other converters in `fromZigbee`, color_colortemp did not
use `postfixWithEndpointName` when populating states. This meant
that multiple endpoints reporting color_mode and color_temp would
clobber each other.

This patch ensures that the proper enpoint suffix is used with
color_mode and color_temp when the device is configured with
multiple endpoints.