Add endpoint name postfix to color cluster state.
The state reported for color cluster do not currently postfix
their keys with the name of the endpoint they are retrieved from.
This causes states reported from multiple endpoints to
overwrite each other.

This patch adds the endpoint name postfixes to the keys of
color cluster attributes.

This is a follow up to #9196 which fixed the issue for color_temp
only and covers other attributes of the color control cluster
in both color_colortemp and in tuya_led_controller.