Add endpoint suffix to level_config converter result
When creating the result state in level_config, the
converter reads in values from the specified endpoint but
does not postfix the key names with endpoint names. This means
that the actual values exposed from multi-endpoint devices
are never used.

This patch updates level_config coverter to postfix its result
with the endpoint name which should allow the values to be
read and handled properly.