fix: improve IKEA light unfreeze logic
This tweaks the implementation of the IKEA unfreeze logic while keeping
the original intent. The previous implementation had some logic flaws
and didn't successfully unfreeze in all scenarios and with all lights
(especially CWS lights).

This new implementation fixes these issues, documents a bit more the
freezing behaviour of IKEA lights, and it's a bit more generic.
Hopefully a bit more maintainable-and future-proof as well.

Fixes: https://github.com/Koenkk/zigbee2mqtt/issues/18574
Fixes: https://github.com/Koenkk/zigbee-herdsman-converters/pull/8637