Add TS0505B_1 transition fixes
This light has two issues:
1. If passing transition = 0, it will behave as if transition = 1.
2. If turning off with a transition, and turning on during the transition, it will turn off
   at the end of the first transition timer, despite order to turn on

Workaround for issue 1: best we can do is set the transition to 0.1 seconds
That is the same thing as is done for TS0505B_2

Workaround for issue 2:
Get the current state of the light after transition time + 0.1s
This won't fix the light's state, but at least it will make us aware that it's off,
allowing user apps to turn it on again if needed.
This could probably be improved by actually turning it on again if necessary.